añadi excepciones personalizadas y corregi el front del login y registro

+++ src/main/java/com/example/onix/models/services/UserService.java
@@ -1,5 +1,9 @@
 package com.example.onix.models.services;
 
+import com.example.onix.exceptions.ConflictException;
+import com.example.onix.exceptions.InvalidCredentialsException;
+import com.example.onix.exceptions.NotFoundException;
+import com.example.onix.mappers.UserMapper;
 import com.example.onix.models.dto.UserDto;
 import com.example.onix.models.entities.User;
 import com.example.onix.models.repositories.UserRepository;
@@ -11,55 +15,77 @@ import java.util.Optional;
 
 @Service
 @AllArgsConstructor
-public class UserService implements IUserService{
+public class UserService implements IUserService {
 
     private final UserRepository userRepository;
+    private final UserMapper userMapper;
 
     @Override
     public List<UserDto> getAllUsers() {
-        return (List<UserDto>) userRepository.findAllDto();
+        List<User> users = (List<User>) userRepository.findAll();
+        if (users.isEmpty()) {
+            throw new NotFoundException("No se encontro ningun usuario");
+        }
+        return userMapper.toDtoList(users);
     }
 
     @Override
-    public void saveUser(User user) {
-        userRepository.save(user);
-    }
-    @Override
-    public void updateUser(UserDto user) {
-        Optional<User> existingUserOpt = userRepository.findById(user.getId());
-        if (existingUserOpt.isEmpty()){
-            throw new RuntimeException("Usuario no encontrado con id: " + user.getId());
+    public UserDto saveUser(User user) {
+
+        if (userRepository.existsUserByEmail(user.getEmail())) {
+            throw new ConflictException("El email ya está registrado");
         }
-            User existingUser = existingUserOpt.get();
-            existingUser.setName(user.getName());
-            existingUser.setEmail(user.getEmail());
-            existingUser.setPassword(user.getPassword());
-            userRepository.save(existingUser);
+
+        return userMapper.toDto(userRepository.save(user));
     }
+
     @Override
-    public UserDto login(UserDto userDto){
+    public UserDto updateUser(UserDto user) {
+        Optional<User> existingUserOpt = userRepository.findById(user.getId());
+        if (existingUserOpt.isEmpty()) {
+            throw new NotFoundException("Usuario no encontrado con id: " + user.getId());
+        }
+        User existingUser = existingUserOpt.get();
+        existingUser.setName(user.getName());
+        existingUser.setEmail(user.getEmail());
+        existingUser.setPassword(user.getPassword());
+
+        return userMapper.toDto(userRepository.save(existingUser));
+    }
+
+    @Override
+    public UserDto login(UserDto userDto) {
         boolean exists = userRepository.existsUserByNameAndPassword(userDto.getName(),
                 userDto.getPassword());
 
-        if (!exists){
-            throw new RuntimeException("Credenciales  incorrectas");
+        if (!exists) {
+            throw new InvalidCredentialsException("Credenciales  incorrectas");
         }
-        return userRepository.findDtoByName(userDto.getName()).orElseThrow(()-> new RuntimeException("Usuario no encontrado"));
-
+        User user = userRepository.findByName(userDto.getName()).orElseThrow(() -> new NotFoundException("Usuario no encontrado"));
+        return userMapper.toDto(user);
     }
 
     @Override
-    public Boolean existUserByEmail(String email){
-        return userRepository.existsUserByEmail(email);
+    public Boolean existUserByEmail(String email) {
+        boolean exists = userRepository.existsUserByEmail(email);
+        if (exists) {
+            throw new ConflictException("El email ingresado ya esta registrado");
+        }
+        return false;
     }
 
     @Override
-    public User getUserById(Long id) {
-        return userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Usuario no encontrado con id: " + id));
+    public UserDto getUserById(Long id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Usuario no encontrado con id: " + id));
+        return userMapper.toDto(user);
     }
+
     @Override
     public void deleteUserById(Long id) {
+        if (!userRepository.existsById(id)){
+            throw new NotFoundException("Usuario no encontrado con id: "+ id);
+        }
         userRepository.deleteById(id);
     }
